Extract SummaryMeasurementTable
This moves the table away from the command that shows it.

Moving the table to its own package required making `PathObjectImageViewers` public so object thumbnails could be displayed.
