+++ qupath-gui-fx/src/main/java/qupath/lib/gui/tools/PathObjectImageViewers.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2022 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2025 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -22,13 +22,14 @@
  */
 
 
-package qupath.lib.gui.commands;
+package qupath.lib.gui.tools;
 
 import java.awt.image.BufferedImage;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.ForkJoinPool;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -49,7 +50,6 @@ import javafx.scene.image.ImageView;
 import javafx.scene.text.TextAlignment;
 import qupath.lib.gui.images.stores.DefaultImageRegionStore;
 import qupath.lib.gui.images.stores.ImageRenderer;
-import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.gui.viewer.OverlayOptions;
 import qupath.lib.gui.viewer.PathObjectPainter;
 import qupath.lib.gui.viewer.QuPathViewer;
@@ -63,104 +63,106 @@ import qupath.lib.regions.RegionRequest;
  * This handles fetching images, and then working with an {@link ImageView} or {@link Canvas}.
  * 
  * @author Pete Bankhead
- * @since v0.4.0 (refactored and updated from earlier code)
+ * @since v0.6.0 (previously non-public)
  */
-class PathObjectImageManagers {
-	
-	
-	private static class PathObjectTableCell<S extends PathObject, T extends PathObject> extends TableCell<S, T> {
+public class PathObjectImageViewers {
 
-		private static final Logger logger = LoggerFactory.getLogger(PathObjectTableCell.class);
-	
-		private PathObjectViewer<Canvas> painter;
-		
-		private Canvas canvas = new Canvas();
-		private double preferredSize = 100;
-		
-//		private static AtomicInteger counter = new AtomicInteger();
-	
-		PathObjectTableCell(PathObjectViewer<Canvas> painter, double padding) {
-			super();
-			logger.trace("Creating new cell ({})", + System.identityHashCode(this));
-			this.setContentDisplay(ContentDisplay.CENTER);
-			this.setAlignment(Pos.CENTER);
-			this.setTextAlignment(TextAlignment.CENTER);
-			this.painter = painter;
-			this.canvas = painter.getNode();
-			canvas.setWidth(preferredSize);
-			canvas.setHeight(preferredSize);
-			canvas.setStyle("-fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.5), 4, 0, 1, 1);");
-			canvas.heightProperty().bind(canvas.widthProperty());
-			
-			tableColumnProperty().addListener((v, o, n) -> {
-				canvas.widthProperty().unbind();
-				if (n != null)
-					canvas.widthProperty().bind(n.widthProperty().subtract(padding*2));
-			});
-	
-			// Update image on width changes
-			// Required to handle the fact that object ROIs may need to be repainted
-			// with updated stroke widths
-			canvas.widthProperty().addListener((v, o, n) -> painter.updateImage(false));
-		}
-	
-	
-		@Override
-		protected void updateItem(T pathObject, boolean empty) {
-			super.updateItem(pathObject, empty);
-			setText(null);
-			if (empty || pathObject == null || pathObject.getROI() == null) {
-				setGraphic(null);
-				return;
-			}
-			setGraphic(canvas);
-			painter.setPathObject(pathObject);
-		}
-		
+	private static ExecutorService sharedPool;
+
+	private static synchronized ExecutorService getSharedPool() {
+		if (sharedPool == null)
+			sharedPool = Executors.newThreadPerTaskExecutor(Thread.ofVirtual().name("pathobject-view", 0L).factory());
+		return sharedPool;
 	}
-	
-	
-	
-	public static <S extends PathObject, T extends PathObject> TableCell<S, T> createTableCell(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, double padding, ExecutorService pool) {
-		var painter = createCanvasPainter(viewer, server, paintObject, pool);
+
+	/**
+	 * Create a table cell to show an object.
+	 * @param viewer the viewer used to get rendering settings and region store
+	 * @param server the server over which the object should be displayed
+	 * @param paintObject the object to display
+	 * @param padding amount of padding to add around the ROI
+	 * @return a new table cell
+	 * @param <S> generic type for the table
+	 * @param <T> generic type for the value in the table cell
+	 */
+	public static <S, T extends PathObject> TableCell<S, T> createTableCell(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, double padding) {
+		var painter = new PathObjectCanvasManager(new Canvas(), viewer, server, paintObject, getSharedPool());
 		return new PathObjectTableCell<>(painter, padding);
 	}
-	
-	public static PathObjectViewer<Canvas> createCanvasPainter(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, ExecutorService pool) {
-		return new PathObjectCanvasManager(new Canvas(), viewer, server, paintObject, pool);
-	}
-	
-	public static PathObjectViewer<ImageView> createImageViewPainter(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, ExecutorService pool) {
-		return new PathObjectImageViewManager(new ImageView(), viewer, server, paintObject, pool);
-	}
-	
-	
-	public abstract static class PathObjectViewer<T extends Node> {
-		
-		private static final Logger logger = LoggerFactory.getLogger(PathObjectViewer.class);
-		
-		private QuPathViewer viewer;
-		private boolean paintObject;
 
-		private ImageServer<BufferedImage> server;
+	/**
+	 * Create an item viewer to display a {@link PathObject} using a canvas.
+	 * @param viewer the viewer used to get rendering settings and region store
+	 * @param server the image to display
+	 * @param paintObject if true, paint the object itself. Otherwise, paint only the corresponding image patch.
+	 * @return the item viewer
+	 */
+	public static ItemViewer<PathObject, Canvas> createCanvasViewer(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject) {
+		return new PathObjectCanvasManager(new Canvas(), viewer, server, paintObject, getSharedPool());
+	}
+
+	/**
+	 * Create an item viewer to display a {@link PathObject} using an {@link ImageView}.
+	 * @param viewer the viewer used to get rendering settings and region store
+	 * @param server the image to display
+	 * @param paintObject if true, paint the object itself. Otherwise, paint only the corresponding image patch.
+	 * @return the item viewer
+	 */
+	public static ItemViewer<PathObject, ImageView> createImageViewer(QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject) {
+		return new PathObjectImageViewManager(new ImageView(), viewer, server, paintObject, getSharedPool());
+	}
+
+	/**
+	 * Interface for managing a node that can display an item.
+	 * <p>
+	 * This is intended to display a {@link PathObject}, but is made generic so in the future it might
+	 * handle something else.
+	 *
+	 * @param <S> generic type for the item to display
+	 * @param <T> generic type for the node that displays the item; typically a {@link Canvas} or an {@link ImageView}.
+	 */
+	public interface ItemViewer<S, T extends Node> {
+
+		/**
+		 * Set the item to display
+		 * @param item the item
+		 */
+		void setItem(S item);
+
+		/**
+		 * Get the node to display
+		 * @return the node
+		 */
+		T getNode();
+
+	}
+	
+	private abstract static class AbstractPathObjectViewer<T extends Node> implements ItemViewer<PathObject, T> {
+		
+		private static final Logger logger = LoggerFactory.getLogger(AbstractPathObjectViewer.class);
+		
+		private final QuPathViewer viewer;
+		private final boolean paintObject;
+
+		private final ImageServer<BufferedImage> server;
 		
 		private PathObject pathObject;
 		
-		private T node;
-		private DoubleProperty widthProperty;
-		private DoubleProperty heightProperty;
+		private final T node;
+		private final DoubleProperty widthProperty;
+		private final DoubleProperty heightProperty;
 
 		private PathObjectThumbnailTask task;
-		private WorkerHandler handler = new WorkerHandler();
+		private final WorkerHandler handler = new WorkerHandler();
 		
 		private int maxCacheSize = 100;
 		
-		private ExecutorService pool;
-		
+		private final ExecutorService pool;
+
 		private WeakChangeListener<Number> numberListenerFalse = new WeakChangeListener<>((v, o, n) -> updateImage(false));
 		private WeakChangeListener<Number> numberListenerTrue = new WeakChangeListener<>((v, o, n) -> updateImage(true));
-		
-		private Map<PathObject, Image> cache = Collections.synchronizedMap(new LinkedHashMap<>() {
+
+		private final Map<PathObject, Image> cache = Collections.synchronizedMap(new LinkedHashMap<>() {
 			
 			private static final long serialVersionUID = 1L;
 
@@ -171,7 +173,7 @@ class PathObjectImageManagers {
 			
 		});
 
-		PathObjectViewer(T node, QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject,
+		AbstractPathObjectViewer(T node, QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject,
 				DoubleProperty widthProperty, DoubleProperty heightProperty, ExecutorService pool) {
 			logger.trace("Creating new cell ({})", + System.identityHashCode(this));
 			this.server = server;
@@ -190,8 +192,9 @@ class PathObjectImageManagers {
 			heightProperty.addListener(numberListenerTrue);
 			node.visibleProperty().addListener(new WeakChangeListener<>((v, o, n) -> updateImage(true)));
 		}
-		
-		public void setPathObject(PathObject pathObject) {
+
+		@Override
+		public void setItem(PathObject pathObject) {
 			if (this.pathObject == pathObject) 
 				return;
 			if (task != null) {
@@ -205,7 +208,8 @@ class PathObjectImageManagers {
 		public PathObject getPathObject() {
 			return pathObject;
 		}
-		
+
+		@Override
 		public T getNode() {
 			return node;
 		}
@@ -272,7 +276,7 @@ class PathObjectImageManagers {
 				
 				pool.submit(task);
 			} catch (Exception e) {
-				logger.error("Problem reading thumbnail for {}: {}", pathObject, e);
+				logger.error("Problem reading thumbnail for {}", pathObject, e);
 			}
 		}
 		
@@ -283,7 +287,7 @@ class PathObjectImageManagers {
 		protected abstract void updateNode(Image image);		
 
 		
-		class WorkerHandler implements EventHandler<WorkerStateEvent> {
+		private class WorkerHandler implements EventHandler<WorkerStateEvent> {
 
 			@Override
 			public void handle(WorkerStateEvent event) {
@@ -309,7 +313,7 @@ class PathObjectImageManagers {
 	}
 	
 	
-	static class PathObjectCanvasManager extends PathObjectViewer<Canvas> {
+	private static class PathObjectCanvasManager extends AbstractPathObjectViewer<Canvas> {
 
 		PathObjectCanvasManager(Canvas canvas, QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, ExecutorService pool) {
 			super(canvas, viewer, server, paintObject,
@@ -327,9 +331,9 @@ class PathObjectImageManagers {
 		}
 		
 	}
-	
-	
-	static class PathObjectImageViewManager extends PathObjectViewer<ImageView> {
+
+
+	private static class PathObjectImageViewManager extends AbstractPathObjectViewer<ImageView> {
 
 		PathObjectImageViewManager(ImageView imageView, QuPathViewer viewer, ImageServer<BufferedImage> server, boolean paintObject, ExecutorService pool) {
 			super(imageView, viewer, server, paintObject,
@@ -352,18 +356,19 @@ class PathObjectImageManagers {
 
 
 
-	static class PathObjectThumbnailTask extends Task<Image> {
+	private static class PathObjectThumbnailTask extends Task<Image> {
 		
 		private static final Logger logger = LoggerFactory.getLogger(PathObjectThumbnailTask.class);
 
-		private DefaultImageRegionStore store;
-		private ImageRenderer renderer;
-		private OverlayOptions options;
+		private final DefaultImageRegionStore store;
+		private final ImageRenderer renderer;
+		private final OverlayOptions options;
 		
-		private ImageServer<BufferedImage> server;
-		private PathObject pathObject;
-		private int width, height;
-		private boolean paintObject;
+		private final ImageServer<BufferedImage> server;
+		private final PathObject pathObject;
+		private final int width;
+        private final int height;
+		private final boolean paintObject;
 
 		PathObjectThumbnailTask(ImageServer<BufferedImage> server, PathObject pathObject, boolean paintObject, int width, int height) {
 			this(server, pathObject, paintObject, width, height, null, null, null);
@@ -427,14 +432,60 @@ class PathObjectImageManagers {
 
 				return SwingFXUtils.toFXImage(img, null);
 			} catch (Exception e) {
-				logger.error("Unable to draw image for " + pathObject, e);
+                logger.error("Unable to draw image for {}", pathObject, e);
 			}
 			return null;
 		}
 
-		
-		
-		
+	}
+
+
+	private static class PathObjectTableCell<S, T extends PathObject> extends TableCell<S, T> {
+
+		private static final Logger logger = LoggerFactory.getLogger(PathObjectTableCell.class);
+
+		private final PathObjectCanvasManager painter;
+
+		private Canvas canvas = new Canvas();
+		private double preferredSize = 100;
+
+		private PathObjectTableCell(PathObjectCanvasManager painter, double padding) {
+			super();
+			logger.trace("Creating new cell ({})", + System.identityHashCode(this));
+			this.setContentDisplay(ContentDisplay.CENTER);
+			this.setAlignment(Pos.CENTER);
+			this.setTextAlignment(TextAlignment.CENTER);
+			this.painter = painter;
+			this.canvas = painter.getNode();
+			canvas.setWidth(preferredSize);
+			canvas.setHeight(preferredSize);
+			canvas.setStyle("-fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.5), 4, 0, 1, 1);");
+			canvas.heightProperty().bind(canvas.widthProperty());
+
+			tableColumnProperty().addListener((v, o, n) -> {
+				canvas.widthProperty().unbind();
+				if (n != null)
+					canvas.widthProperty().bind(n.widthProperty().subtract(padding*2));
+			});
+
+			// Update image on width changes
+			// Required to handle the fact that object ROIs may need to be repainted
+			// with updated stroke widths
+			canvas.widthProperty().addListener((v, o, n) -> painter.updateImage(false));
+		}
+
+
+		@Override
+		protected void updateItem(T pathObject, boolean empty) {
+			super.updateItem(pathObject, empty);
+			setText(null);
+			if (empty || pathObject == null || pathObject.getROI() == null) {
+				setGraphic(null);
+				return;
+			}
+			setGraphic(canvas);
+			painter.setItem(pathObject);
+		}
 
 	}
